ticket: 7085 subject: Better short/long descs in gss_display_mech_attr
RFC 5587 defines an "arc name" and a "purpose" for each mechanism
attribute.  gss_display_mech_attr was returning the purpose (which
isn't very short) in short_desc and a null buffer in long_desc (which
isn't necessarily conformant to the spec).  Instead, output the arc
name as the short_desc and the purpose as the long_desc.  This is more
useful and also more consistent with Heimdal, which outputs the arc
name as short_desc and either the purpose string or an empty string
for long_desc.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@25668 dc483132-0cff-0310-8789-dd5450dbe970