Remove user_appl section in RST docs
The user_appl directory contained documentation for krb5_appl programs
which have largely been subsumed by ssh, for ksu (essentially
duplicating the ksu man page), and for ssh (a stub).  Get rid of the
whole thing for now; we can add an ssh section back into the user
guide when we have more written about it.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@25766 dc483132-0cff-0310-8789-dd5450dbe970